rasppi-requirements.txt
opencv-python-headless
gpiozero
RPi.GPIO